Fix empty fields when browser opens app in issue #2

+++ app/src/main/java/xyz/hisname/fireflyiii/repository/models/auth/AuthModel.kt
@@ -2,7 +2,7 @@ package xyz.hisname.fireflyiii.repository.models.auth
 
 data class AuthModel(
         val token_type: String,
-        val expires_in: Int,
+        val expires_in: Long,
         val access_token: String,
         val refresh_token: String
 )
